- fix atomnotimplemented tendrl-bug-id: Tendrl/specifications#37
